chore: refactor hooks into same olog package; move AppInfo hook into app pkg; update README
